added lib.rs which allows the import from other crates

+++ src/plt.rs
@@ -1,13 +1,13 @@
 use plotters::prelude::*;
 
 // prepare_vec returns a 2d vector suitable for plotting and also min, max values of input vector
-fn prepare_vec(vals: Vec<f64>) -> (Vec<(f32, f64)>, f64, f64) {
+fn prepare_vec(vals: Vec<f64>) -> (Vec<(f64, f64)>, f64, f64) {
     let mut out = vec![(0.0, 0.0); vals.len()];
     let mut min = vals[0];
     let mut max = vals[0];
 
     for i in 0..vals.len() {
-        out[i] = (i as f32, vals[i]);
+        out[i] = (i as f64, vals[i]);
         if vals[i] > max {
             max = vals[i]
         } else if vals[i] < min {
@@ -28,7 +28,7 @@ pub fn plt(vals: Vec<f64>, filename: &str) -> Result<(), Box<dyn std::error::Err
         .margin(5)
         .x_label_area_size(30)
         .y_label_area_size(30)
-        .build_ranged(0f32..vec2d.len() as f32, min..max)?;
+        .build_ranged(0f64..vec2d.len() as f64, min..max)?;
 
     chart.configure_mesh().draw()?;
